major update in progress, changing to handle misaligned atom numbers

+++ quiver.sh
@@ -21,6 +21,9 @@ if [ ! -f $2 ]; then
     exit 1
 fi
 
+# pre cleanup
+rm -f *.q1 *.q2 *.q3 *.qout \ *
+
 # prepare QUIVER input files
 awk -f quiver_prep.awk $1 $2
 if [ $? -ne 0 ]; then
@@ -32,18 +35,27 @@ fi
 cp gs.q1 temp.q1
 cp gs.q2 temp.q2
 rm -f temp.qout
-./quiver.exe &> /dev/null
+./quiver.exe
+if [ $? -ne 0 ]; then
+    echo "Something went wrong when running QUIVER on the ground state."
+    exit 1
 cp temp.qout gs.qout
 
 # run QUIVER on transition state
 cp ts.q1 temp.q1
 cp ts.q2 temp.q2
 rm -f temp.qout
-./quiver.exe &> /dev/null
+./quiver.exe
+if [ $? -ne 0 ]; then
+    echo "Something went wrong when running QUIVER on the ground state."
+    exit 1
 cp temp.qout ts.qout
 
 # run the analysis
 awk -f quiver_analysis.awk gs.q3 gs.qout ts.qout
+if [ $? -ne 0 ]; then
+    echo "Something went wrong when performing the analysis."
+    exit 1
 
 # clean up
 rm -f *.q1 *.q2 *.q3 *.qout \ *
